// src/app/models/enums.ts
export enum WorkEnum {
  Hibryd = 'Hybrid work',
  Flexible = 'Flexible work',
  Remote = 'Remote work',
  Home = 'Work from Home',
  OnSite = 'On-site Work',
  Shift = 'Shift Work',
  Mobile = 'Mobile Work',
  Mixed = 'Mixed-mode Work'
}

export enum SkillEnum
{
  CSharp = 'C#',
  WPF = 'WPF',
  WF = 'Windows Form',
  Angular = 'Angular',
  HTML = 'HTML',
  CSS = 'CSS',
  TS = 'TypeScript'
}

export enum PatternEnum{
  MVVM = 'MVVM',
  CQRS = 'CQRS',
  AGILE = 'Agile',
  SCRUM = 'SCRUM',
  DDD = 'Domain Driven Design (DDD)'
}